ml/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
xgboost
joblib
pillow
tensorflow==2.20.1
